# openapi/target/classes/openapi.yml
openapi: 3.0.3
info:
  title: API oPROsita
  description: API REST para gestión de grupos, usuarios, contenido y simulacros
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
...
paths:
...
  /alumnos/{id}:
    delete:
      tags:
        - alumnos
      operationId: deleteAlumno
      summary: Eliminar un alumno
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Alumno eliminado correctamente
  /alumnos/{alumnoId}/contenido:
    get:
      tags:
        - alumnos
        - contenido
      summary: Obtener contenido de un alumno
      operationId: getContenidoByAlumno
      parameters:
        - name: alumnoId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Lista de contenido
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ContenidoItem'

    post:
      tags:
        - alumnos
        - contenido
      operationId: addContenidoToAlumno
      summary: Añadir contenido al alumno (texto + tipo + PDF)
      parameters:
        - in: path
          name: alumnoId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                texto:
                  type: string
                tipoContenido:
                  type: string
                  enum: [ temas, programacion, practico ]
                mes:
                  type: string
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Contenido añadido
  /alumnos/{alumnoId}/contenido/{contenidoId}:
    delete:
      tags:
        - alumnos
        - contenido
      operationId: deleteContenidoFromAlumno
      summary: Eliminar un contenido del alumno
      parameters:
        - in: path
          name: alumnoId
          required: true
          schema:
            type: integer
        - in: path
          name: contenidoId
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Contenido eliminado correctamente
...
  /alumnos/{alumnoId}/grupos:
    get:
      tags:
        - alumnos
        - grupos
      summary: Obtener grupos de un alumno
      operationId: getGruposByAlumno
      parameters:
        - name: alumnoId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Lista de grupos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Grupo'
...
  /archivos/upload:
    post:
      tags:
        - archivos
      operationId: uploadArchivo
      summary: Subir archivo
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Archivo subido correctamente
  /archivos/{id}/download:
    get:
      tags:
        - archivos
      operationId: downloadArchivo
      summary: Descargar archivo por ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Archivo descargado
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          description: Archivo no encontrado
...
components:
  schemas:
    Grupo:
      type: object
      properties:
        id:
          type: integer
        nombre:
          type: string
        meses:
          type: array
          items:
            $ref: '#/components/schemas/Mes'
    Mes:
      type: object
      properties:
        id:
          type: integer
        nombre:
          type: string
...
    ContenidoItem:
      type: object
      properties:
        id:
          type: integer
        texto:
          type: string
        tipoContenido:
          type: string
          enum: [temas, programacion, practico]
        alumnoId:
          type: integer
          nullable: true
        grupoId:
          type: integer
        mes:
          type: string
        archivoId:
          type: integer
          nullable: true
...